requirements.txt
asgiref==3.5.0
Django==3.1.4
pytz==2022.1
sqlparse==0.4.2
djangorestframework